+++ test/typescript/hello-world/index.ts
@@ -58,10 +58,8 @@ broker.loadService(path.join(__dirname, "posts.service.ts"));
 		await broker.call("v2.posts.list", { limit: 10, sort: "title" });
 		await broker.call("greeter.hello");
 		const res = await broker.call("greeter.welcome", { name: "Typescript" });
-		broker.logger.info("");
-		broker.logger.info("Result: ", res);
-		broker.logger.info("");
-		if (res != "Welcome, TYPESCRIPT!")
+		broker.logger.info(`Result: ${res}`);
+		if (res != "Welcome, TYPESCRIPT")
 			throw new Error("Result is mismatch!");
 		else
 			await broker.stop();